adding a minor safety check for dead link removal

+++ scala-default.nix
@@ -123,7 +123,9 @@ in { scalaEnv = buildEnv {
   # TODO: better filter, use ammonite script?:
   postBuild = ''
     # we remove a dead symbolic link, which currently causes nix to break:
-    rm $out/lib/libunwind-generic.a
+    if [ ! -e $out/lib/libunwind-generic.a ] ; then
+      rm $out/lib/libunwind-generic.a
+    fi
   # for f in $(ls -d $out/bin/* | grep "idea"); do
   #   sed -i '/IDEA_JDK/d' $f
   #   wrapProgram $f \
